web/requirements.txt
aiobotocore==2.3.4
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aioitertools==0.11.0
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
beautifulsoup4==4.12.3
boto3==1.21.21
botocore==1.24.21
bs4==0.0.2
certifi==2024.7.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.2.8
django-ckeditor==6.7.0
django-dump-load-utf8==0.0.4
django-environ==0.11.2
django-import-export==2.8.0
django-js-asset==2.2.0
django-recaptcha==4.0.0
django-storages==1.14.4
djangorestframework==3.16.1
et-xmlfile==1.1.0
frozenlist==1.4.1
fsspec==2022.5.0
idna==3.7
influxdb-client==1.36.1
jmespath==1.0.1
multidict==6.0.5
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pandas==1.4.2
Pillow==10.0.0
plotly==5.23.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
reactivex==4.0.4
s3fs==2022.5.0
s3transfer==0.5.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tablib==3.6.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==1.26.19
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
fpdf==1.7.2
kaleido==0.2.1
requests==2.32.3
kaleido==0.2.1
celery==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==1.2.0
python-crontab==3.1.0
redis==4.5.1
mysqlclient
mysql-connector-python